requirements.txt
python-dotenv
nest_asyncio
streamlit
langchain
langchain-community
langchain-anthropic
langchain-fireworks
langchain-openai
openai
langchain-google-vertexai
langchain-google-community
langchain-cohere
langchain-mistralai
nltk
Markdown
pytesseract
pypdf
PyPDF2
PyMuPDF
unstructured[all-docs]
html2text
playwright
beautifulsoup4
duckduckgo-search
faiss-cpu
huggingface-hub
accelerate
torch
sentence-transformers
sentencepiece
transformers
langdetect
pycountry
gpt-researcher
flash-attn